Add custom toolbar with logo image
