+++ xamarin/Samples/AndorraTelecom-Android/Activities/MainActivity.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using Android.App;
+using Android.OS;
+using Android.Widget;
+
+namespace AndorraTelecomAndroid
+{
+    [Activity(MainLauncher = true, Icon = "@mipmap/icon")]
+    public class MainActivity : Activity
+    {
+        protected override void OnCreate(Bundle savedInstanceState)
+        {
+            base.OnCreate(savedInstanceState);
+            SetContentView(Resource.Layout.Main);
+
+            SetCustomActionBar();
+        }
+
+        void SetCustomActionBar()
+        {
+            Toolbar Toolbar = (Toolbar)FindViewById(Resource.Id.toolbar);
+            SetActionBar(Toolbar);
+            ActionBar.SetDisplayShowTitleEnabled(false);
+            ActionBar.SetIcon(Resource.Drawable.logo);
+        }
+    }
+}
