+++ xamarin/Samples/AndorraTelecom-Android/Properties/AndroidManifest.xml
@@ -1,5 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="1" android:versionName="1.0" package="com.liferay.mobile.screens.AndorraTelecom_Android">
 	<uses-sdk android:minSdkVersion="10" />
-	<application android:allowBackup="true" android:icon="@mipmap/icon" android:label="@string/app_name" android:theme="@style/Theme.AppCompat.Light"></application>
+	<application
+        android:allowBackup="true"
+        android:icon="@mipmap/icon"
+        android:label="@string/app_name"
+        android:theme="@style/AndorraTelecomTheme">
+    </application>
 </manifest>
