+++ xamarin/Samples/AndorraTelecom-Android/Resources/layout/Main.axml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Hello!" />
+    <include
+        android:id="@+id/toolbar"
+        layout="@layout/toolbar" />
 </LinearLayout>
