+++ xamarin/Samples/AndorraTelecom-Android/Resources/values/Colors.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<resources>
+    <color name="colorAccent_andorra">@color/colorAccent</color>
+    <color name="colorPrimary_andorra">@color/colorPrimary</color>
+    <color name="colorPrimaryDark_andorra">@color/colorPrimaryDark</color>
+    <color name="textColorPrimary_andorra">@color/textColorPrimary</color>
+    
+    <color name="colorPrimary">#C6005F</color>
+    <color name="colorPrimaryDark">#790045</color>
+    <color name="colorAccent">#790045</color>
+    <color name="textColorPrimary">#FFFFFF</color>
+</resources>
