+++ xamarin/Samples/AndorraTelecom-Android/Resources/values/Strings.xml
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <resources>
-	<string name="hello">Hello World, Click Me!</string>
-	<string name="app_name">AndorraTelecom-Android</string>
+	<string name="app_name">Andorra Telecom</string>
 </resources>
